+++ src/pyomsdk/resources/blockstore_resource.py
@@ -70,13 +70,13 @@ Expected response body
 Setting these tags limits which backup jobs this blockstore can process. If omitted, this blockstore can only process backup jobs for projects that do not use labels to filter their jobs.
         """
 
-        load_factor: Optional[float] = Field(None, serialization_alias="loadFactor")
+        load_factor: Optional[int] = Field(None, serialization_alias="loadFactor")
         """Optional. A positive, non-zero integer that expresses how much backup work this snapshot store should perform compared to another snapshot store. This option is needed only if more than one snapshot store is in use.
 
 To learn more about Load Factor, see Edit One Existing Blockstore.
         """
 
-        max_capacity_gb: Optional[float] = Field(None, serialization_alias="maxCapacityGB")
+        max_capacity_gb: Optional[int] = Field(None, serialization_alias="maxCapacityGB")
         """Optional. The blockstore’s expected maximum available capacity for use in the OUTSIDE_SPACE_USED_THRESHOLD alert. maxCapacityGB does not enforce any limitation on the size of the backing database.
         """
 
@@ -206,11 +206,11 @@ Some API clients cannot access the HTTP response headers or status code. To reme
 For endpoints that return a list of results, the content object is an envelope. Ops Manager adds the status field to the response body.
         """
 
-        items_per_page: Optional[float] = Field(100.0, serialization_alias="itemsPerPage")
+        items_per_page: Optional[int] = Field(100, serialization_alias="itemsPerPage")
         """Number of items to return per page, up to a maximum of 500.
         """
 
-        page_num: Optional[float] = Field(1.0, serialization_alias="pageNum")
+        page_num: Optional[int] = Field(1, serialization_alias="pageNum")
         """One-based integer that returns a subsection of results.
         """
 
@@ -366,13 +366,13 @@ Expected response body
 Setting these tags limits which backup jobs this blockstore can process. If omitted, this blockstore can only process backup jobs for projects that do not use labels to filter their jobs.
         """
 
-        load_factor: Optional[float] = Field(None, serialization_alias="loadFactor")
+        load_factor: Optional[int] = Field(None, serialization_alias="loadFactor")
         """Optional. A positive, non-zero integer that expresses how much backup work this snapshot store should perform compared to another snapshot store. This option is needed only if more than one snapshot store is in use.
 
 To learn more about Load Factor, see Edit One Existing Blockstore.
         """
 
-        max_capacity_gb: Optional[float] = Field(None, serialization_alias="maxCapacityGB")
+        max_capacity_gb: Optional[int] = Field(None, serialization_alias="maxCapacityGB")
         """Optional. The blockstore’s expected maximum available capacity for use in the OUTSIDE_SPACE_USED_THRESHOLD alert. maxCapacityGB does not enforce any limitation on the size of the backing database.
         """
