+++ src/pyomsdk/resources/global_events_resource.py
@@ -53,7 +53,7 @@ To review the types of events that generate alerts, see Alert Types.
 For a complete list of events included in the Ops Manager audit log, see Audit Events.
         """
 
-        items_per_page: Optional[float] = Field(100.0, serialization_alias="itemsPerPage")
+        items_per_page: Optional[int] = Field(100, serialization_alias="itemsPerPage")
         """Number of items to return per page, up to a maximum of 500.
         """
 
@@ -65,7 +65,7 @@ For a complete list of events included in the Ops Manager audit log, see Audit E
         """Return only events for which the created date is greater than or equal to the specified Timestamp in ISO 8601 date and time format in UTC.
         """
 
-        page_num: Optional[float] = Field(1.0, serialization_alias="pageNum")
+        page_num: Optional[int] = Field(1, serialization_alias="pageNum")
         """Page number (1-index based).
         """
